Rename loginactivity -> signinactivity add authentication system bookmarkactivity prototype
