+++ app/src/main/java/com/example/myongsubway/SignUpFragment.java
@@ -7,6 +7,7 @@ import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentTransaction;
 
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -18,19 +19,29 @@ import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.gms.tasks.Task;
+import com.google.firebase.auth.AuthResult;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.FirebaseFirestore;
+import com.google.firebase.firestore.auth.User;
+
+import org.jetbrains.annotations.NotNull;
 
 import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.concurrent.Executor;
+
+import static android.content.ContentValues.TAG;
 
 /**
  * 회원가입 프래그먼트
- * 이메일과 비밀번호를 통해 문서ID를 만들고 생성합니다.
+ * 이메일과 비밀번호를 통해 문서 ID를 만들고 생성합니다.
  */
 public class SignUpFragment extends Fragment {
 
+    private FirebaseAuth mAuth;
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
     private static final String ARG_PARAM1 = "param1";
     private static final String ARG_PARAM2 = "param2";
@@ -76,38 +87,42 @@ public class SignUpFragment extends Fragment {
 
         //파이어스토어에 접근하기 위한 객체
         final FirebaseFirestore db = FirebaseFirestore.getInstance();
-
+        // Initialize Firebase Auth
+        mAuth = FirebaseAuth.getInstance();
         DocumentReference docRef = db.collection("users").document(getUserData());
-
-        // 문서를 만듭니다.(회원가입)
-        ArrayList<String> list;
-        HashMap<String, ArrayList<String>> map;
-
-        list = new ArrayList<String>();
-        map = new HashMap<String, ArrayList<String>>();
 
         complete.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                // 새 문서와 새 아이디 생성
-                db.collection("users").document(getUserData())
-                        .set(map)
-                        .addOnSuccessListener(new OnSuccessListener<Void>() {
-                            @Override
-                            public void onSuccess(Void aVoid) {
-                                System.out.println("DocumentSnapshot successfully written!");
-                            }
-                        })
-                        .addOnFailureListener(new OnFailureListener() {
-                            @Override
-                            public void onFailure(@NonNull Exception e) {
-                                System.out.println("Error writing document");
-                            }
-                        });
-
-                // 버튼 클릭 후 마지막으로 프래그먼트 종료
-                FragmentManager fragmentManager = getActivity().getSupportFragmentManager();
-                fragmentManager.beginTransaction().remove(SignUpFragment.this).commit();
+                mAuth.createUserWithEmailAndPassword(getEmail(), getPassword())
+                        .addOnCompleteListener(new OnCompleteListener<AuthResult>() {
+                    @Override
+                    public void onComplete(@NonNull @NotNull Task<AuthResult> task) {
+                        if (task.isSuccessful()) {
+                            ArrayList<String> list = new ArrayList<String>();
+                            HashMap<String, ArrayList<String>> map = new HashMap<String, ArrayList<String>>();
+                            // Add a new document with a generated ID
+                            db.collection("subwayData").document(mAuth.getUid())
+                                    .set(map)
+                                    .addOnSuccessListener(new OnSuccessListener<Void>() {
+                                        @Override
+                                        public void onSuccess(Void aVoid) {
+                                            // 버튼 클릭 후 성공하면 마지막으로 프래그먼트 종료
+                                            FragmentManager fragmentManager = getActivity().getSupportFragmentManager();
+                                            fragmentManager.beginTransaction().remove(SignUpFragment.this).commit();
+                                        }
+                                    })
+                                    .addOnFailureListener(new OnFailureListener() {
+                                        @Override
+                                        public void onFailure(@NonNull Exception e) {
+                                            Log.w(TAG, "Error writing document:failure", task.getException());
+                                        }
+                                    });
+                        } else {
+                            Log.w(TAG, "createUserWithEmail:failure", task.getException());
+                        }
+                    }
+                });
             }
         });
         return v;
@@ -116,4 +131,5 @@ public class SignUpFragment extends Fragment {
     String getEmail(){ return email.getText().toString(); }
     String getPassword(){ return password.getText().toString(); }
     String getUserData(){return getEmail() +"_" + getPassword();}
+
 }
