+++ app/src/main/res/layout/activity_sign_in.xml
@@ -2,11 +2,11 @@
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/login_constraintlayout_login"
+    android:id="@+id/signin_constraintlayout_signin"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".LoginActivity">
+    tools:context=".SignInActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -14,7 +14,7 @@
         android:orientation="vertical">
 
         <TextView
-            android:id="@+id/login_textview_login"
+            android:id="@+id/signin_textview_signin"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_gravity="center"
@@ -23,7 +23,7 @@
             android:textSize="24dp" />
 
         <ImageView
-            android:id="@+id/login_imageView_image"
+            android:id="@+id/signin_imageView_image"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_gravity="center"
@@ -31,38 +31,38 @@
             android:src="@mipmap/ic_launcher" />
 
         <TextView
-            android:id="@+id/login_edittext_email"
+            android:id="@+id/signin_edittext_email"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:text="이메일 입력" />
 
         <EditText
-            android:id="@+id/login_edittext_emailinput"
+            android:id="@+id/signin_edittext_emailinput"
             android:layout_width="match_parent"
             android:layout_height="50dp"
             android:hint="user@gmail.com" />
 
         <TextView
-            android:id="@+id/login_textview_pawssword"
+            android:id="@+id/signin_textview_pawssword"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:text="비밀번호 입력" />
 
         <EditText
-            android:id="@+id/login_edittext_passwordinput"
+            android:id="@+id/signin_edittext_passwordinput"
             android:layout_width="match_parent"
             android:layout_height="50dp"
             android:inputType="textPassword"
             android:hint="****" />
 
         <Button
-            android:id="@+id/login_button_login"
+            android:id="@+id/signin_button_signin"
             android:layout_width="match_parent"
             android:layout_height="52dp"
             android:text="로그인" />
 
         <TextView
-            android:id="@+id/login_textview_or"
+            android:id="@+id/signin_textview_or"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_gravity="center"
@@ -71,7 +71,7 @@
             android:textSize="14dp" />
 
         <Button
-            android:id="@+id/login_button_withoutlogin"
+            android:id="@+id/signin_button_withoutsignin"
             android:layout_width="match_parent"
             android:layout_height="52dp"
             android:text="로그인 없이 시작" />
@@ -83,7 +83,7 @@
             android:weightSum="10">
 
             <TextView
-                android:id="@+id/login_textview_findpassword"
+                android:id="@+id/signin_textview_findpassword"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:layout_weight="5"
@@ -92,7 +92,7 @@
                 android:textSize="18dp" />
 
             <TextView
-                android:id="@+id/login_textview_signup"
+                android:id="@+id/signin_textview_signup"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:layout_weight="5"
